requirements.txt
Django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
gunicorn
whitenoise
psycopg2-binary
